apps: clean open posix warning
warning: format ‘%ld’ expects argument of type ‘long int’, but argument 4 has type ‘time_t’ {aka ‘unsigned int’} [-Wformat=]
   85 |     printf("PASS - slept %ds%dns ~= %lds%ldns\n",
      |                                     ~~^
      |                                       |
      |                                       long int
      |                                     %d
   86 |            slepts, sleptns,
   87 |            tssleepfor.tv_sec, tssleepfor.tv_nsec);
      |            ~~~~~~~~~~~~~~~~~
      |                      |
      |                      time_t {aka unsigned int}

warning: passing argument 3 of ‘pthread_attr_getstack’ from incompatible pointer type [-Wincompatible-pointer-types]
   47 |  rc = pthread_attr_getstack(&attr, &stack_addr, &stack_size);
      |                                                 ^~~~~~~~~~~
      |                                                 |
      |                                                 size_t * {aka unsigned int *}

warning: integer overflow in expression of type ‘long int’ results in ‘-2147483648’ [-Woverflow]
   31 |   structure_and_field = LONG_MAX+1;  \

warning: format not a string literal and no format arguments [-Wformat-security]
   57 |   printf(ts);
      |   ^~~~~~

warning: control reaches end of non-void function [-Wreturn-type]
  375 | }
      | ^

warning: ‘dummy’ defined but not used [-Wunused-variable]
    7 | static sigset_t dummy;
      |                 ^~~~~

warning: variable ‘dummyvar’ set but not used [-Wunused-but-set-variable]
   17 |  sigtimedwait_test dummyvar;
      |                    ^~~~~~~~

warning: ‘test_mq_timedsend_prototype’ defined but not used [-Wunused-function]
   11 | static void test_mq_timedsend_prototype(void)
      |             ^~~~~~~~~~~~~~~~~~~~~~~~~~~

warning: function declaration isn’t a prototype [-Wstrict-prototypes]
   17 | static void dummy_func()
      |             ^~~~~~~~~~

error: passing argument 3 of ‘pthread_create’ from incompatible pointer type [-Werror=incompatible-pointer-types]
   53 |  if (pthread_create(&new_th, &attr, a_thread_func, NULL) != 0) {
      |                                     ^~~~~~~~~~~~~
      |                                     |
      |                                     void * (void)

+++ testcases/open_posix_testsuite/conformance/interfaces/clock_gettime/3-1.c
@@ -58,7 +58,7 @@ int main(void)
 	}
 
 	printf("Test FAILED - ts1=%ld,ts2=%ld,ts3=%ld,ts4=%ld\n",
-	       ts1.tv_sec, ts2.tv_sec, ts3.tv_sec, ts4.tv_sec);
+	       (long int)ts1.tv_sec, (long int)ts2.tv_sec, (long int)ts3.tv_sec, (long int)ts4.tv_sec);
 	return PTS_FAIL;
 #else
 	printf("CLOCK_MONOTONIC unsupported\n");
